trunk: Change member variables of ADFun (except Rec) to end in underbar.
This was done with an automatic editing script. Because Rec is such a 
short name, it was excluded form the editing pass. For each of the
other ADFun member variables, the following sed commands were used to
mofify the files below:
	s/\([^a-zA-Z_]\)old_name\([^a-zA-Z_]\)/\1new_name\2/g
	s/\([^a-zA-Z_]\)old_name$/\1new_name/g
where old_name and new_name are the old and new names for the variables.



git-svn-id: https://projects.coin-or.org/svn/CppAD/trunk@1165 0e562018-b8fb-0310-a933-ecf073c7c197