trunk: Change get_element to next_element (do not check all element values).
This is in preparation templating the type of vector of sets and having
two options for the sparsity calculations. Both vector of sets will have 
the same API and in the case where sets are implemented as std::set, 
it will be more efficient to not check every possible element value.

makefile.in: set corresponding makefile.am version (missing in previous commit).



git-svn-id: https://projects.coin-or.org/svn/CppAD/trunk@1527 0e562018-b8fb-0310-a933-ecf073c7c197